Try to make inkjs less chatty on the console.
InkJS writes to the console object in a few places. While this is fine when running a single story, it can get bad on a service that potentially runs many many stories in parallel. Especially when this service reloads the story for every choice the user makes (in order to be able to act stateless) and there are console warnings when loading the story file. Such a service will want to control what to log and how itself.

For one such warning that was written to the console when loading the story, I changed it to use the existing mechanism to collect warnings for the story instead.
Assert calls now also don't log to the console anymore but simple put the assertion message in the Error that they already throw.

Some console logs remain, for example in SimpleJson.ts - but they'd be a bit trickier to fix. I will change them as well should they become a problem.